Enable only parts of the tpnm feed to be parsed

+++ nrod/nrod-tpnm/src/main/java/uk/trainwatch/nrod/tpnm/tool/TPNMImport.java
@@ -10,41 +10,39 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.UncheckedIOException;
 import java.nio.file.Path;
-import java.sql.CallableStatement;
 import java.sql.Connection;
-import java.sql.PreparedStatement;
 import java.sql.SQLException;
-import java.sql.Timestamp;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Objects;
-import java.util.function.Consumer;
+import java.util.Set;
 import java.util.function.Function;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import java.util.stream.Stream;
-import javax.xml.datatype.XMLGregorianCalendar;
 import javax.xml.parsers.ParserConfigurationException;
 import org.apache.commons.cli.CommandLine;
 import org.kohsuke.MetaInfServices;
-import uk.trainwatch.nrod.tpnm.model.Directed;
 import uk.trainwatch.nrod.tpnm.model.Node;
-import uk.trainwatch.nrod.tpnm.model.Point;
+import uk.trainwatch.nrod.tpnm.model.Projectinfo;
+import uk.trainwatch.nrod.tpnm.model.Projecttype;
+import uk.trainwatch.nrod.tpnm.model.Schedversiontype;
 import uk.trainwatch.nrod.tpnm.model.Signal;
 import uk.trainwatch.nrod.tpnm.model.Station;
 import uk.trainwatch.nrod.tpnm.model.Stationcategorydesc;
 import uk.trainwatch.nrod.tpnm.model.Stationtypedesc;
-import uk.trainwatch.nrod.tpnm.model.Track;
 import uk.trainwatch.nrod.tpnm.model.Trackcategorydesc;
 import uk.trainwatch.nrod.tpnm.model.Uiccodedesc;
-import uk.trainwatch.nrod.tpnm.model.Way;
-import uk.trainwatch.util.ParserUtils;
+import uk.trainwatch.nrod.tpnm.model.Version;
+import uk.trainwatch.util.Consumers;
+import static uk.trainwatch.util.Consumers.sink;
 import uk.trainwatch.util.Router;
 import uk.trainwatch.util.app.DBUtility;
 import uk.trainwatch.util.app.Utility;
 import uk.trainwatch.util.counter.RateMonitor;
 import uk.trainwatch.util.sql.SQL;
+import uk.trainwatch.util.sql.SQLBiConsumer;
 import uk.trainwatch.util.sql.SQLConsumer;
-import uk.trainwatch.util.sql.UncheckedSQLException;
 import uk.trainwatch.util.xml.XMLStreamParserBuilder;
 
 /**
@@ -58,14 +56,27 @@ public class TPNMImport
 
     protected static final Logger LOG = Logger.getLogger( TPNMImport.class.getName() );
     private static final String SCHEMA = "tpnm";
+
+    private final RateMonitor<org.w3c.dom.Node> monitor;
+    private final Function<org.w3c.dom.Node, Object> unmarshaller;
+
     private List<Path> cifFiles;
     private boolean fullImport;
+    private boolean textImport, stationImport, nodeImport, signalImport;
 
     public TPNMImport()
     {
         super();
         getOptions().
-                addOption( null, "full", false, "Full rather than Incremental import" );
+                addOption( null, "full", false, "Full rather than Incremental import" ).
+                addOption( null, "text", false, "Import text only" ).
+                addOption( null, "station", false, "Import stations" ).
+                addOption( null, "node", false, "Import nodes" ).
+                addOption( null, "signal", false, "Import signals" );
+
+        monitor = RateMonitor.log( LOG, "Records" );
+
+        unmarshaller = new TPNMUnmarshaller();
     }
 
     @Override
@@ -75,6 +86,16 @@ public class TPNMImport
         super.parseArgs( cmd );
 
         fullImport = cmd.hasOption( "full" );
+
+        textImport = fullImport || cmd.hasOption( "text" );
+        stationImport = fullImport || cmd.hasOption( "station" );
+        nodeImport = fullImport || cmd.hasOption( "node" );
+        signalImport = fullImport || cmd.hasOption( "signal" );
+
+        if( !(fullImport || textImport || stationImport || nodeImport || signalImport) ) {
+            LOG.log( Level.SEVERE, "No options set" );
+            return false;
+        }
 
         // The first one will be the CIF name
         cifFiles = Utility.getArgFileList( cmd );
@@ -86,22 +107,21 @@ public class TPNMImport
     protected void initDB( Connection con )
             throws SQLException
     {
-        Stream.of(
-                "security_way",
-                "securitysection",
-                "signal",
-                //
-                "waypoint",
-                "node",
-                "way",
-                //
-                "track",
-                "station",
-                "stationtypedesc",
-                "stationcategorydesc",
-                "trackcategorydesc",
-                "uiccodedesc"
-        ).forEach( k -> SQL.deleteTable( con, SCHEMA, k ) );
+        if( fullImport ) {
+            LOG.log( Level.INFO, "Clearing down database" );
+            Stream.of(
+                    "security_way",
+                    "securitysection",
+                    "signal",
+                    //
+                    "waypoint",
+                    "node",
+                    "way",
+                    //
+                    "track",
+                    "station"
+            ).forEach( k -> SQL.deleteTable( con, SCHEMA, k ) );
+        }
 
     }
 
@@ -122,82 +142,105 @@ public class TPNMImport
         } );
     }
 
-    private Consumer<Object> lastRoute;
-
     private void importFile( Connection con, Path path )
             throws SQLException,
                    IOException,
                    ParserConfigurationException
     {
-        RateMonitor<org.w3c.dom.Node> monitor = RateMonitor.log( LOG, "Records" );
+        monitor.reset();
 
-        Router<Class<?>, Object> router = Router.createClassRouter().
-                addSQL( Uiccodedesc.class, v -> importUicCodeDesk( con, (Uiccodedesc) v ) ).
-                addSQL( Trackcategorydesc.class, new TextImporter<>( con,
-                                                                     SCHEMA, "trackcategorydesc",
-                                                                     Trackcategorydesc::getId,
-                                                                     Trackcategorydesc::getText,
-                                                                     Trackcategorydesc::getLastmodified ) ).
-                addSQL( Stationcategorydesc.class, new TextImporter<>( con,
-                                                                       SCHEMA, "stationcategorydesc",
-                                                                       Stationcategorydesc::getId,
-                                                                       Stationcategorydesc::getText,
-                                                                       Stationcategorydesc::getLastmodified ) ).
-                addSQL( Stationtypedesc.class, new TextImporter<>( con,
-                                                                   SCHEMA, "stationtypedesc",
-                                                                   Stationtypedesc::getId,
-                                                                   Stationtypedesc::getText,
-                                                                   Stationtypedesc::getLastmodified ) ).
-                addSQL( Station.class, new StationImporter( con ) ).
-                addSQL( Node.class, new NodeImporter( con ) ).
-                addSQL( Signal.class, new SignalImporter( con ) );
+        // Sink for disabled supported types
+        SQLConsumer<Object> sink = Consumers.sqlSink();
 
-        lastRoute = null;
+        Router<Class<?>, Object> router = Router.createClassRouter();
+        Set<Class<?>> unsupported = new HashSet<>();
+
+        // Commit between root object types but not any that are sinks
+        router.setKeyChangeConsumer( SQLBiConsumer.guard( ( k, v ) -> {
+            if( router.isRouted( k ) ) {
+                LOG.log( Level.INFO, () -> "Committing " + k.getSimpleName() );
+                con.commit();
+            }
+        } ) );
+
+        // Log anything not supported then sink the rest
+        router.setSink( o -> {
+            Class<?> c = o.getClass();
+            if( !unsupported.contains( c ) ) {
+                LOG.log( Level.INFO, () -> "Unsupported " + o.getClass().getSimpleName() );
+                unsupported.add( c );
+            }
+        } );
+
+        // These are always unsupported or we have no use for them
+        unsupported.add( Version.class );
+        unsupported.add( Projecttype.class );
+        unsupported.add( Schedversiontype.class );
+        unsupported.add( Projectinfo.class );
+
+        // Text imports
+        if( textImport ) {
+            router.addSQL( Uiccodedesc.class, new UiccodedescImporter( con ) ).
+                    addSQL( Trackcategorydesc.class, new TextImporter<>( con,
+                                                                         SCHEMA, "trackcategorydesc",
+                                                                         Trackcategorydesc::getId,
+                                                                         Trackcategorydesc::getText,
+                                                                         Trackcategorydesc::getLastmodified ) ).
+                    addSQL( Stationcategorydesc.class, new TextImporter<>( con,
+                                                                           SCHEMA, "stationcategorydesc",
+                                                                           Stationcategorydesc::getId,
+                                                                           Stationcategorydesc::getText,
+                                                                           Stationcategorydesc::getLastmodified ) ).
+                    addSQL( Stationtypedesc.class, new TextImporter<>( con,
+                                                                       SCHEMA, "stationtypedesc",
+                                                                       Stationtypedesc::getId,
+                                                                       Stationtypedesc::getText,
+                                                                       Stationtypedesc::getLastmodified ) );
+        }
+        else {
+            // Disabled so sink them
+            unsupported.add( Uiccodedesc.class );
+            unsupported.add( Trackcategorydesc.class );
+            unsupported.add( Stationcategorydesc.class );
+            unsupported.add( Stationtypedesc.class );
+        }
+
+        // The custom importers
+        if( stationImport ) {
+            router.addSQL( Station.class, new StationImporter( con ) );
+        }
+        else {
+            unsupported.add( Station.class );
+        }
+
+        if( nodeImport ) {
+            router.addSQL( Node.class, new NodeImporter( con ) );
+        }
+        else {
+            unsupported.add( Node.class );
+        }
+
+        if( signalImport ) {
+            router.addSQL( Signal.class, new SignalImporter( con ) );
+        }
+        else {
+            unsupported.add( Signal.class );
+        }
 
         LOG.log( Level.INFO, () -> "Parsing " + path );
         try( InputStream is = new FileInputStream( path.toFile() ) ) {
             new XMLStreamParserBuilder().
                     setInputStream( is ).
                     setTriggerDepth( 2 ).
-                    //setAppendDepth( 2 ).
                     createStream().
                     peek( monitor ).
-                    map( new TPNMUnmarshaller() ).
+                    map( unmarshaller ).
                     filter( Objects::nonNull ).
-                    //limit( 20 ).
-                    forEach( o -> {
-                        Consumer<Object> c = router.apply( o );
-                        if( c != lastRoute && lastRoute != null ) {
-                            try {
-                                LOG.log( Level.INFO, "Committing work" );
-                                con.commit();
-                            }
-                            catch( SQLException ex ) {
-                                throw new UncheckedSQLException( ex );
-                            }
-                        }
-                        lastRoute = c;
-
-                        c.accept( o );
-                    } );
+                    forEach( router );
         }
         finally {
             LOG.log( Level.INFO, () -> "Complete " + monitor );
         }
-    }
-
-    private void importUicCodeDesk( Connection con, Uiccodedesc v )
-            throws SQLException
-    {
-        try( PreparedStatement ps = SQL.prepare( con,
-                                                 "INSERT INTO " + SCHEMA + ".uiccodedesc VALUES (?,?,?,?)",
-                                                 v.getId(),
-                                                 v.getText(),
-                                                 new Timestamp( v.getLastmodified().toGregorianCalendar().toInstant().getEpochSecond() ),
-                                                 v.getVanillatext() ) ) {
-            ps.executeUpdate();
-        }
-
     }
 
 }
